+++ src/app/Playground/HåvardSamples/HåvardPage.xaml.cs
@@ -3,6 +3,7 @@ using System.Windows.Input;
 using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.Chips;
 using DIPS.Mobile.UI.Components.Navigation.FloatingNavigationButton;
+using SkiaSharp.Extended.UI.Controls;
 
 namespace Playground.HåvardSamples;
 
@@ -13,8 +14,13 @@ public partial class HåvardPage
         InitializeComponent();
     }
 
-    public ICommand SearchCommand { get; } = new Command<string>(s =>
+    public ICommand SearchCommand => new Command<string>(async s =>
     {
-
+        
     });
+
+    private void OnPropertyChanged(object sender, PropertyChangedEventArgs e)
+    {
+        
+    }
 }