+++ src/library/DIPS.Mobile.UI/API/Library/DUI.cs
@@ -1,5 +1,6 @@
 using DIPS.Mobile.UI.Components.BottomSheets;
 using DIPS.Mobile.UI.Components.Navigation.FloatingNavigationButton;
+using SkiaSharp.Extended.UI.Controls.Themes;
 
 namespace DIPS.Mobile.UI.API.Library
 {
@@ -22,6 +23,21 @@ namespace DIPS.Mobile.UI.API.Library
             }
             
             RemovePlatformSpecificViewsLocatedOnTopOfPage();
+        }
+        
+        public static void EnsureSkLottieResourcesAdded()
+        {
+            // try register with the current app
+            var merged = Application.Current?.Resources?.MergedDictionaries;
+            if (merged == null)
+            {
+                return;
+            }
+
+            if (merged.All(dic => dic.GetType() != typeof(SKLottieViewResources)))
+            {
+                merged.Add(new SKLottieViewResources());
+            }
         }
 
         private static partial void RemovePlatformSpecificViewsLocatedOnTopOfPage();